changed UI and code, details in desc
- changed recipeId to int
- added DAO queries to count and list recipe lists (Main/Favorites/MyRecipes)
- Favorites now show recipes that are added to favorites, added filters to Main and MyRecipe activities as well (correct data is shown)
- Recipe activity shows name, will be finished later
- added RecycleView UI for all 3 activities
- empty MyRecipe activity now shows a button to AddRecipe activity
- empty Favorites activity shows text

+++ app/src/main/java/fi/group6/dessertrecipeapp/classes/Recipe.java
@@ -26,7 +26,7 @@ import androidx.room.PrimaryKey;
 public class Recipe {
 
     @PrimaryKey(autoGenerate = true)
-    public long recipeId;
+    public int recipeId;
 
     @ColumnInfo(name = "name")
     public String name;
